+++ server/src/models/auth.rs
@@ -6,12 +6,12 @@ use crate::db::get_postgres;
 #[derive(FromRow, Serialize, Deserialize)]
 pub struct Auth {
     id: String,
-    account: i64,
-    expire: i64,
+    account: i32,
+    expire: i32,
 }
 
 impl Auth {
-    pub async fn insert_auth_info(id: &str, account: i64, expire: i64) -> Result<(), sqlx::Error> {
+    pub async fn insert_auth_info(id: &str, account: i32, expire: i32) -> Result<(), sqlx::Error> {
         let pool = get_postgres();
 
         // delete overdue data
@@ -34,7 +34,7 @@ impl Auth {
         Ok(())
     }
 
-    pub async fn check_auth_info(id: &str, account: i64) -> bool {
+    pub async fn check_auth_info(id: &str, account: i32) -> bool {
         let r = sqlx::query_as::<_, Auth>("select * from auth where id = $1 and account = $2;")
             .bind(id)
             .bind(account)
